Big refactoring. Moved things to other files.
